Added a module with various delegate decorators to aiida_vasp/utils/delegates.py
