+++ aiida_vasp/parsers/vasp.py
@@ -1,6 +1,5 @@
 #encoding: utf-8
 """AiiDA Parser for a aiida_vasp.VaspCalculation"""
-from functools import update_wrapper
 import numpy as np
 from aiida.orm import DataFactory
 
@@ -107,38 +106,6 @@ PARSABLE_FILES = {
         'status': 'Unknown'
     },
 }
-
-
-def delegate():
-    """
-    Get a decorator adding attributes to add or remove functions to a list of functions.
-    When the decorated function is called, all functions in the list will be called.
-    """
-
-    def decorator(meth):
-        """Decorate a class method to delegate kwargs."""
-
-        meth.listeners = []
-
-        def add_listener(func):
-            meth.listeners.append(func)
-
-        def remove_listener(func):
-            if func in meth.listeners:
-                meth.listeners.remove(func)
-
-        setattr(meth, 'add_listener', add_listener)
-        setattr(meth, 'remove_listener', remove_listener)
-
-        def wrapper(*args, **kwargs):
-            for func in meth.listeners:
-                func(*args, **kwargs)
-            meth(*args, **kwargs)
-
-        update_wrapper(wrapper, meth)
-        return wrapper
-
-    return decorator
 
 
 class VaspParser(BaseParser):
